+++ assets/js/blocks/cart-checkout/cart/full-cart/index.js
@@ -8,12 +8,12 @@ import {
 	TotalsCoupon,
 	TotalsDiscount,
 	TotalsFooterItem,
+	TotalsShipping,
 } from '@woocommerce/base-components/cart-checkout';
 import {
 	Subtotal,
 	TotalsFees,
 	TotalsTaxes,
-	TotalsShipping,
 	ExperimentalOrderMeta,
 	getCurrencyFromPriceResponse,
 } from '@woocommerce/blocks-checkout';
