+++ packages/checkout/shipping/shipping-rates-control/index.js
@@ -1,2 +1,1 @@
-export { default as ShippingRatesControl } from './shipping-rates-control';
 export { default as ShippingRatesControlPackage } from './package';
