+++ packages/checkout/totals/index.js
@@ -1,5 +1,4 @@
 export { default as TotalsItem } from './item';
 export { default as Subtotal } from './subtotal';
 export { default as TotalsTaxes } from './taxes';
-export { default as TotalsShipping } from './shipping';
 export { default as TotalsFees } from './fees';
